+++ example/tx-client-demo-a/src/main/java/com/codingapi/lcn/tx/demo/a/DemoAApplication.java
@@ -1,0 +1,19 @@
+package com.codingapi.lcn.tx.demo.a;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author lorne
+ * @date 2018/8/30
+ * @description
+ */
+@SpringBootApplication
+public class DemoAApplication {
+
+
+    public static void main(String[] args) {
+        SpringApplication.run(DemoAApplication.class,args);
+    }
+
+}
