+++ example/tx-client-demo-a/src/main/java/com/codingapi/lcn/tx/demo/a/service/DemoService.java
@@ -1,0 +1,14 @@
+package com.codingapi.lcn.tx.demo.a.service;
+
+import com.codingapi.lcn.tx.demo.a.domain.User;
+
+/**
+ * @author lorne
+ * @date 2018/8/30
+ * @description
+ */
+public interface DemoService {
+
+
+    boolean saveUser(User user);
+}
