+++ example/tx-client-demo-start/src/main/resources/bootstrap.properties
@@ -1,0 +1,17 @@
+server.port=8099
+
+spring.application.name=start
+
+
+spring.datasource.driver-class-name = com.mysql.jdbc.Driver
+spring.datasource.url= jdbc:mysql://127.0.0.1:3306/m_test?characterEncoding=UTF-8
+spring.datasource.username= root
+spring.datasource.password= root
+
+
+spring.cloud.consul.host=127.0.0.1
+spring.cloud.consul.port=8500
+spring.cloud.consul.discovery.prefer-ip-address=true
+
+tx.client.jdbc-aspect-order=1
+tx.client.transaction-aspect-order=1
