--- example/tx-client-demo/src/main/java/com/codingapi/lcn/tx/demo/DemoApplication.java
@@ -1,19 +1,0 @@
-package com.codingapi.lcn.tx.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * @author lorne
- * @date 2018/8/30
- * @description
- */
-@SpringBootApplication
-public class DemoApplication {
-
-
-    public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class,args);
-    }
-
-}
