--- example/tx-client-demo/src/main/java/com/codingapi/lcn/tx/demo/domain/User.java
@@ -1,28 +1,0 @@
-package com.codingapi.lcn.tx.demo.domain;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * @author lorne
- * @date 2018/8/30
- * @description
- */
-
-@Data
-@AllArgsConstructor
-@Slf4j
-@NoArgsConstructor
-public class User {
-
-
-    private Long id;
-
-
-    private String name;
-
-
-}
