--- example/tx-client-demo/src/main/java/com/codingapi/lcn/tx/demo/service/DemoService.java
@@ -1,14 +1,0 @@
-package com.codingapi.lcn.tx.demo.service;
-
-import com.codingapi.lcn.tx.demo.domain.User;
-
-/**
- * @author lorne
- * @date 2018/8/30
- * @description
- */
-public interface DemoService {
-
-
-    boolean saveUser(User user);
-}
