+++ tx-client-api/src/main/java/com/codingapi/lcn/tx/api/TxClientApiConfiguration.java
@@ -1,0 +1,16 @@
+package com.codingapi.lcn.tx.api;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author lorne
+ * @date 2018/8/31
+ * @description
+ */
+@Configuration
+@ComponentScan
+public class TxClientApiConfiguration {
+
+
+}
