+++ tx-client-api/src/main/java/com/codingapi/lcn/tx/api/sleuth/ISleuthTransactionApi.java
@@ -1,0 +1,16 @@
+package com.codingapi.lcn.tx.api.sleuth;
+
+/**
+ * @author lorne
+ * @date 2018/8/31
+ * @description
+ */
+public interface ISleuthTransactionApi {
+
+
+    /**
+     * 当地是否是开始模块
+     * @return
+     */
+    boolean isStart();
+}
