+++ tx-client-api/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.codingapi.lcn.tx.api.TxClientApiConfiguration
